write & read works when using single process, bugs need to be fixed for multi-process run
