+++ source/adios2/engine/bp4/BP4FileReader.h
@@ -52,6 +52,8 @@ private:
     format::BP4Deserializer m_BP4Deserializer;
     transportman::TransportMan m_FileManager;
     transportman::TransportMan m_SubFileManager;
+    /*Lipeng*/
+    transportman::TransportMan m_FileMetadataIndexManager;
 
     /** used for per-step reads, TODO: to be moved to BP4Deserializer */
     size_t m_CurrentStep = 0;
