+++ source/adios2/toolkit/format/bp4/BP4Serializer.cpp
@@ -305,16 +305,23 @@ void BP4Serializer::AggregateCollectiveMetadata(MPI_Comm comm,
 
     auto &position = bufferSTL.m_Position;
 
-    const uint64_t pgIndexStart =
+    /*Lipeng*/
+    //const uint64_t pgIndexStart =
+    //    inMetadataBuffer ? position : position + bufferSTL.m_AbsolutePosition;
+    m_MetadataSet.pgIndexStart =
         inMetadataBuffer ? position : position + bufferSTL.m_AbsolutePosition;
     AggregateIndex(m_MetadataSet.PGIndex, m_MetadataSet.DataPGCount, comm,
                    bufferSTL);
 
-    const uint64_t variablesIndexStart =
+    //const uint64_t variablesIndexStart =
+    //    inMetadataBuffer ? position : position + bufferSTL.m_AbsolutePosition;
+    m_MetadataSet.varIndexStart =
         inMetadataBuffer ? position : position + bufferSTL.m_AbsolutePosition;
     AggregateMergeIndex(m_MetadataSet.VarsIndices, comm, bufferSTL);
 
-    const uint64_t attributesIndexStart =
+    //const uint64_t attributesIndexStart =
+    //    inMetadataBuffer ? position : position + bufferSTL.m_AbsolutePosition;
+    m_MetadataSet.attrIndexStart =
         inMetadataBuffer ? position : position + bufferSTL.m_AbsolutePosition;
     AggregateMergeIndex(m_MetadataSet.AttributesIndices, comm, bufferSTL, true);
 
@@ -322,9 +329,10 @@ void BP4Serializer::AggregateCollectiveMetadata(MPI_Comm comm,
     MPI_Comm_rank(comm, &rank);
     if (rank == 0)
     {
-        PutMinifooter(pgIndexStart, variablesIndexStart, attributesIndexStart,
-                      bufferSTL.m_Buffer, bufferSTL.m_Position,
-                      inMetadataBuffer);
+        /*Lipeng*/
+        //PutMinifooter(pgIndexStart, variablesIndexStart, attributesIndexStart,
+        //              bufferSTL.m_Buffer, bufferSTL.m_Position,
+        //              inMetadataBuffer);
 
         if (inMetadataBuffer)
         {
