--- source/adios2/toolkit/format/bp4/backup/BP4.h
@@ -1,17 +1,0 @@
-/*
- * Distributed under the OSI-approved Apache License, Version 2.0.  See
- * accompanying file Copyright.txt for details.
- *
- * BP4.h : public header for the BP4 format
- *
- *  Created on: Aug 3, 2018
- *      Author: Lipeng Wan wanl@ornl.gov
- */
-
-#ifndef ADIOS2_TOOLKIT_FORMAT_BP4_BP4_H_
-#define ADIOS2_TOOLKIT_FORMAT_BP4_BP4_H_
-
-#include "adios2/toolkit/format/bp4/BP4Deserializer.h"
-#include "adios2/toolkit/format/bp4/BP4Serializer.h"
-
-#endif /* ADIOS2_TOOLKIT_FORMAT_BP4_BP4_H_ */
